// src/app/model/Customer.ts
export class Customer {

    customerId: number = 0;
    customerFirstName: string = '';
    customerLastName: string = '';
    email: string = '';
    phone: string = '';
    address: string = '';
    state: string = '';
    city: string = '';
    username: string = '';
    password: string = '';
    agentId: number = 0;
    status: boolean = true;
    queryCount: number = 0;
    documentsCount: number = 0;
    policiesCount: number = 0;
    dob: Date = new Date();
    userId: number = 0;
}